Implement Azure pipeline
It was a test code to see if Mac issues were related to GitHub actions
and will not repro on the Azure pipelines.

The test has shown that the issue is with the CI Mac image that
has Mac OS 12 instead of Mac OS 13 that fails the AOT code when
it throws exceptions

Keeping this file just in case if it is useful in future.